Initial version of my RCP script.
This one just starts logging as soon as the engine is running and
the gps has fix, then stops logging with a delay after the engine
stops. This should mean that you don't get a ton of log files when
the engine won't idle and constantly stalls. Yes that happens often
enough to our car to be worth considering. Welcom to LeMons.