added gear ratio computation.
The supra has a speed sensor at the transmission output shaft, so
I'm just using that to compute current gear ratio directly. I plan
to later use this information to figure out what gear I'm in. This
is the way easier way to do things relative to using GPS to guess the
current gear ratio...