Added gear number to the gear ratio calculation!
Now we can log what gear we're in! Sweet!

There are some known bugs:
This currently has no idea if you're holding the clutch in or not,
so as you coast down from a higher gear, you'll appear to briefly
enter all the lower gears on your way, possibly including reverse.
I should figure out how to filter for that in some useful way...