first pass at making the fan work via RCP
I ditched the fan sensor thing and I'm going with RCP now. I hope
this can eventually lead to more useful fan behavior, but I was also
running out of room for the sensor now that there is also a coolant
pressure sensor!

you can also see a bit of the beginnings of the warning stuff here.
I should have waited to commit that. I know, I'm being lazy, but hey,
this is my side project!