first pass at a warning method.
I'm writing this from the hotel room near the track, so it isn't
tested yet, but it looks like it migth work to me, so I'm going to
call it good and catch some sleep before the race tomorrow.
this just sets some simple thresholds for the warnings. I know they
will bounce a lot near the actual event, but I don't think that's
actually a terrible thing in this case, as long as the relay survives.
I might put a bit of sticky logic in there at some point if it
seems like it is needed.

in the future this should be more complex so it can handle things
like RPM dependent oil pressure, temperature sensor differences,
and initial startup.