after some profiling, here is a version that is quite a bit faster
this does use the size sorting hack, but also turns off all the
processing until the very end to avoid processing stuff over and over
and over again. That's a good thing. I also saved a full minute by
using xrange instead of range. I guess that really does matter when
you're calling it 3.4 million times...

on a full race's data, this currently takes 217 seconds on my laptop.
That's much slower than I'd like, but not nearly as bad as it was with
totally uncontrolled multiple processing. I think it is time to call it
a night.

next up: write the merged thing out so I don't have to keep merging it
all the time, although I think at this point most of the time is in the
parsing and in the object creation. according to cProfile.run, more than
2 minutes is spent in side of object.__new__. That sound bad...